requirements.txt
flask
python-telegram-bot==20.7
requests
python-dotenv
gunicorn